refactor(vats): separate core behaviors, manifests, and other parts
