+++ packages/vats/decentral-core-config.json
@@ -2,94 +2,7 @@
   "bootstrap": "bootstrap",
   "vats": {
     "bootstrap": {
-      "sourceSpec": "./src/core/boot.js",
-      "parameters": {
-        "bootstrapManifest": {
-          "connectVattpWithMailbox": {
-            "vatPowers": {
-              "D": true
-            },
-            "vats": {
-              "vattp": true
-            },
-            "devices": {
-              "mailbox": true
-            }
-          },
-          "makeVatsFromBundles": {
-            "vats": {
-              "vatAdmin": true
-            },
-            "devices": {
-              "vatAdmin": true
-            },
-            "produce": {
-              "vatAdminSvc": true,
-              "loadVat": true
-            }
-          },
-          "buildZoe": {
-            "consume": {
-              "agoricNames": true,
-              "nameAdmins": true,
-              "vatAdminSvc": true,
-              "loadVat": true,
-              "client": true
-            },
-            "produce": {
-              "zoe": true,
-              "feeMintAccess": true
-            }
-          },
-          "makeBoard": {
-            "consume": {
-              "loadVat": true,
-              "client": true
-            },
-            "produce": {
-              "board": true
-            }
-          },
-          "makeAddressNameHubs": {
-            "consume": {
-              "client": true
-            },
-            "produce": {
-              "agoricNames": true,
-              "agoricNamesAdmin": true,
-              "nameAdmins": true
-            }
-          },
-          "startTimerService": {
-            "devices": {
-              "timer": true
-            },
-            "vats": {
-              "timer": true
-            },
-            "produce": {
-              "chainTimerService": true
-            }
-          },
-          "makeClientBanks": {
-            "consume": {
-              "loadVat": true,
-              "client": true,
-              "bridgeManager": true
-            },
-            "produce": {
-              "bankManager": true
-            }
-          },
-          "makeBLDKit": {
-            "consume": {
-              "agoricNames": true,
-              "bankManager": true,
-              "nameAdmins": true
-            }
-          }
-        }
-      }
+      "sourceSpec": "./src/core/boot.js"
     }
   },
   "bundles": {
