+++ packages/vats/src/core/utils.js
@@ -1,0 +1,106 @@
+// @ts-check
+import { E, Far } from '@agoric/far';
+import { AssetKind } from '@agoric/ertp';
+import { makeNotifierKit } from '@agoric/notifier';
+
+const { entries, fromEntries } = Object;
+
+// TODO: phase out ./issuers.js
+export const CENTRAL_ISSUER_NAME = 'RUN';
+
+// We reserve these keys in name hubs.
+export const shared = harden({
+  // issuer, brand nameAdmins
+  assets: {
+    BLD: 'Agoric staking token',
+    RUN: 'Agoric RUN currency',
+    Attestation: 'Agoric lien attestation',
+  },
+  // installation, instance nameAdmins
+  contract: {
+    contractGovernor: 'contract governor',
+    committee: 'committee electorate',
+    Attestation: 'Agoric lien attestation',
+    getRUN: 'getRUN',
+  },
+});
+
+/** @type { FeeIssuerConfig } */
+export const feeIssuerConfig = {
+  name: CENTRAL_ISSUER_NAME,
+  assetKind: AssetKind.NAT,
+  displayInfo: { decimalPlaces: 6, assetKind: AssetKind.NAT },
+};
+
+/**
+ * Wire up a remote between the comms vat and vattp.
+ *
+ * @param {string} addr
+ * @param {{ vats: { vattp: VattpVat, comms: CommsVatRoot }}} powers
+ */
+export const addRemote = async (addr, { vats: { comms, vattp } }) => {
+  const { transmitter, setReceiver } = await E(vattp).addRemote(addr);
+  await E(comms).addRemote(addr, transmitter, setReceiver);
+};
+harden(addRemote);
+
+/**
+ * @param { string } addr
+ * @param {{
+ *   vats: {
+ *     vattp: VattpVat,
+ *     comms: CommsVatRoot,
+ *   },
+ *   produce: { client: Producer<ClientConfig>, chainBundler: Producer<ChainBundler> },
+ * }} powers
+ */
+export const installClientEgress = async (
+  addr,
+  { vats: { comms, vattp }, produce: { client, chainBundler } },
+) => {
+  const PROVISIONER_INDEX = 1;
+
+  let bundle = harden({});
+  const { notifier, updater } = makeNotifierKit({ bundle });
+
+  /** @type {ChainBundler} */
+  const chainProvider = Far('chainProvider', {
+    getChainBundle: () =>
+      notifier.getUpdateSince().then(({ value: { bundle: b } }) => b),
+    getChainConfigNotifier: () => notifier,
+  });
+  chainBundler.resolve(chainProvider);
+
+  await addRemote(addr, { vats: { comms, vattp } });
+  await E(comms).addEgress(addr, PROVISIONER_INDEX, chainProvider);
+
+  const callProperties = (obj, ...args) =>
+    fromEntries(entries(obj).map(([k, fn]) => [k, fn(...args)]));
+
+  client.resolve(
+    harden({
+      assignBundle: newPropertyMakers => {
+        const newProperties = callProperties(newPropertyMakers, addr);
+        bundle = { ...bundle, ...newProperties };
+        updater.updateState({ bundle });
+      },
+    }),
+  );
+};
+harden(installClientEgress);
+
+/**
+ * @param {string[]} edges
+ * @param {ERef<NameHub>} agoricNames
+ * @param {ERef<Store<NameHub, NameAdmin>>} nameAdmins
+ * @returns {Promise<NameAdmin[]>}
+ */
+export const collectNameAdmins = (edges, agoricNames, nameAdmins) => {
+  return Promise.all(
+    edges.map(async edge => {
+      const hub = /** @type {NameHub} */ (await E(agoricNames).lookup(edge));
+      return E(nameAdmins).get(hub);
+    }),
+  );
+};
+harden(collectNameAdmins);
